fix living dex status bug -- now refresh does not reset the living dex toggle
