+++ src/Variables/globalVariables.ts
@@ -1,2 +1,6 @@
 export const POKEMON_LIMIT = 898;
 export const MAX_POKEMON = 898;
+
+export const SIMPLE_KEY = "allSimplePokemon";
+export const LIVING_DEX_KEY = "livingDex";
+export const LIVING_DEX_STATUS_KEY = "livingDexStatus"
